updated with working code
This is the code that was submitted... and it works!
